+++ components/templates/LoginPage.vue
@@ -1,0 +1,14 @@
+<template>
+  <LoginPanel />
+</template>
+
+<script>
+import { defineComponent } from '@nuxtjs/composition-api'
+import LoginPanel from '~/components/organisms/Login/LoginPanel.vue'
+
+export default defineComponent({
+  components: {
+    LoginPanel
+  }
+})
+</script>
